xrefr:main/1: Halt with a non-zero exit status if warnings
Before this change, the `xrefr` escript could exit with 1 if there was
an issue with the command line flags. Now, it exits with 1 if there are
warnings reported. To not confuse this situation with a bad usage, the
former exit status was changed to 64, which is EX_USAGE in sysexits(3).

This helps when `xrefr` is executed from another script or Makefile
because it's easy to know when there are warnings reported without
parsing the output.